refactor(audio): migrate core types from MopidyTrack to Track
