+++ yoyopy/audio/__init__.py
@@ -7,16 +7,19 @@ Provides audio playback, volume control, and device management.
 from yoyopy.audio.history import RecentTrackEntry, RecentTrackHistoryStore
 from yoyopy.audio.local_service import LocalLibraryItem, LocalMusicService
 from yoyopy.audio.manager import AudioManager, AudioDevice
-from yoyopy.audio.mopidy_client import MopidyClient, MopidyTrack, MopidyPlaylist
+from yoyopy.audio.music import MusicBackend, MockMusicBackend, MpvBackend, Track, Playlist, MusicConfig
 
 __all__ = [
-    'AudioManager',
     'AudioDevice',
+    'AudioManager',
     'LocalLibraryItem',
     'LocalMusicService',
+    'MockMusicBackend',
+    'MpvBackend',
+    'MusicBackend',
+    'MusicConfig',
+    'Playlist',
     'RecentTrackEntry',
     'RecentTrackHistoryStore',
-    'MopidyClient',
-    'MopidyTrack',
-    'MopidyPlaylist',
+    'Track',
 ]
